requirements.txt
Django==1.8.4
djangorestframework==3.2.3
requests==2.7.0
wsgiref==0.1.2
